+++ src/test/java/nextstep/subway/domain/SubwayMapTest.java
@@ -8,7 +8,6 @@ import com.google.common.collect.Lists;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
-import org.jgrapht.GraphPath;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -80,26 +79,7 @@ class SubwayMapTest {
      */
     @BeforeEach
     void setUp() {
-        교대역 = createStation(1L, "교대역");
-        강남역 = createStation(2L, "강남역");
-        양재역 = createStation(3L, "양재역");
-        남부터미널역 = createStation(4L, "남부터미널역");
-        삼성역 = createStation(5L, "삼성역");
-        대치역 = createStation(6L, "대치역");
-
-        신분당선 = Line.of("신분당선", "red", 2000
-            , "0520", "2340", 20);
-        이호선 = Line.of("2호선", "red", 900
-            , "0500", "2330", 10);
-        삼호선 = Line.of("3호선", "red", 1100
-            , "0530", "2300", 5);
-
-        신분당선.addSection(강남역, 양재역, 9, 2);
-        이호선.addSection(교대역, 강남역, 8, 3);
-        이호선.addSection(강남역, 삼성역, 8, 3);
-        삼호선.addSection(교대역, 남부터미널역, 6, 5);
-        삼호선.addSection(남부터미널역, 양재역, 6, 5);
-        삼호선.addSection(양재역, 대치역, 2, 5);
+        createSubwayMap();
     }
 
     @DisplayName("KShortest 알고리즘을 사용하여 가장 최단거리 경로를 조회한다.")
@@ -108,11 +88,11 @@ class SubwayMapTest {
         // given
         List<Line> lines = Lists.newArrayList(신분당선, 이호선, 삼호선);
         SubwayMap subwayMap = new SubwayMap(lines);
-        List<GraphPath<Station, SectionEdge>> allPaths = subwayMap.findAllKShortestPaths(
+        AllKShortestPaths allPaths = subwayMap.findAllKShortestPaths(
             강남역, 남부터미널역, PathType.DISTANCE);
 
         // when
-        ShortestPaths shortest = subwayMap.findShortest(allPaths, START_TIME);
+        ShortestPaths shortest = allPaths.getShortestPathsFrom(START_TIME);
 
         // then
         assertAll(
@@ -129,11 +109,11 @@ class SubwayMapTest {
         // given
         List<Line> lines = Lists.newArrayList(신분당선, 이호선, 삼호선);
         SubwayMap subwayMap = new SubwayMap(lines);
-        List<GraphPath<Station, SectionEdge>> allPaths = subwayMap.findAllKShortestPaths(
+        AllKShortestPaths allPaths = subwayMap.findAllKShortestPaths(
             강남역, 남부터미널역, PathType.DISTANCE);
 
         // when
-        ShortestPaths shortest = subwayMap.findShortest(allPaths, START_TIME);
+        ShortestPaths shortest = allPaths.getShortestPathsFrom(START_TIME);
 
         // then
         assertAll(
@@ -142,91 +122,6 @@ class SubwayMapTest {
             () -> assertThat(shortest.getShortestDistance()).isEqualTo(14),
             () -> assertThat(shortest.getShortestDurationArrivalTime()).isEqualTo(SHORTEST_DURATION_ARRIVAL_TIME)
         );
-    }
-
-    @DisplayName("단일 구간에서 상행선 방향의 열차 탑승 시간을 확인한다.")
-    @Test
-    void findTrainTime_singleSection_upDirection_Test() {
-        // given
-        List<Line> lines = Lists.newArrayList(신분당선, 이호선, 삼호선);
-        SubwayMap subwayMap = new SubwayMap(lines);
-
-        // when
-        LocalDateTime trainTime = subwayMap.findTrainTime(
-            new Section(신분당선, 양재역, 강남역, 9, 2), convertStringToDateTime("202202200605"));
-
-        // then
-        assertThat(trainTime).isEqualTo(convertStringToDateTime("202202200620"));
-    }
-
-    @DisplayName("단일 구간에서 하행선 방향의 열차 탑승 시간을 확인한다.")
-    @Test
-    void findTrainTime_singleSection_downDirection_Test() {
-        // given
-        List<Line> lines = Lists.newArrayList(신분당선, 이호선, 삼호선);
-        SubwayMap subwayMap = new SubwayMap(lines);
-
-        // when
-        LocalDateTime trainTime = subwayMap.findTrainTime(
-            new Section(신분당선, 강남역, 양재역, 9, 2), convertStringToDateTime("202202200600"));
-
-        // then
-        assertThat(trainTime).isEqualTo(convertStringToDateTime("202202200600"));
-    }
-
-    @DisplayName("전체 노선에서 특정 구간의 상행선 열차 탑승 시간을 확인한다.")
-    @Test
-    void findTrainTime_UpDirection_Test() {
-        // given
-        List<Line> lines = Lists.newArrayList(신분당선, 이호선, 삼호선);
-        SubwayMap subwayMap = new SubwayMap(lines);
-
-        LocalDateTime trainTime = subwayMap.findTrainTime(
-            new Section(이호선, 강남역, 교대역, 8, 3), convertStringToDateTime(START_TIME));
-
-        assertThat(trainTime).isEqualTo(LocalDateTime.of(2022, 02, 20, 06, 03));
-    }
-
-    @DisplayName("전체 노선에서 특정 구간의 하행선 열차 탑승 시간을 확인한다.")
-    @Test
-    void findTrainTime_DownDirection_Test() {
-        // given
-        List<Line> lines = Lists.newArrayList(신분당선, 이호선, 삼호선);
-        SubwayMap subwayMap = new SubwayMap(lines);
-
-        LocalDateTime trainTime = subwayMap.findTrainTime(
-            new Section(이호선, 교대역, 강남역, 8, 3), convertStringToDateTime(START_TIME));
-
-        assertThat(trainTime).isEqualTo(LocalDateTime.of(2022, 02, 20, 06, 00));
-    }
-
-    @DisplayName("경로의 구간이 어느 방향행인지 확인해보니 상행이다.")
-    @Test
-    void findPathUpDirectionTest() {
-        // given
-        List<Line> lines = Lists.newArrayList(신분당선, 이호선, 삼호선);
-        SubwayMap subwayMap = new SubwayMap(lines);
-
-        // when
-        PathDirection direction = subwayMap.findPathDirection(
-            new Section(이호선, 강남역, 교대역, 8, 3));
-
-        // then
-        assertThat(direction).isEqualTo(PathDirection.UP);
-    }
-
-    @DisplayName("경로의 구간이 어느 방향행인지 확인해보니 하행이다.")
-    @Test
-    void findPathDownDirectionTest() {
-        // given
-        List<Line> lines = Lists.newArrayList(신분당선, 이호선, 삼호선);
-        SubwayMap subwayMap = new SubwayMap(lines);
-
-        // when
-        PathDirection direction = subwayMap.findPathDirection(new Section(삼호선, 남부터미널역, 양재역, 8, 3));
-
-        // then
-        assertThat(direction).isEqualTo(PathDirection.DOWN);
     }
 
     @DisplayName("LocalDateTime 몇 분 차이가 나는지 확인한다 (양수)")
@@ -309,6 +204,29 @@ class SubwayMapTest {
             () -> assertThat(path.getStations())
                 .containsExactlyElementsOf(Lists.newArrayList(남부터미널역, 교대역, 강남역))
         );
+    }
+
+    private void createSubwayMap() {
+        교대역 = createStation(1L, "교대역");
+        강남역 = createStation(2L, "강남역");
+        양재역 = createStation(3L, "양재역");
+        남부터미널역 = createStation(4L, "남부터미널역");
+        삼성역 = createStation(5L, "삼성역");
+        대치역 = createStation(6L, "대치역");
+
+        신분당선 = Line.of("신분당선", "red", 2000
+            , "0520", "2340", 20);
+        이호선 = Line.of("2호선", "red", 900
+            , "0500", "2330", 10);
+        삼호선 = Line.of("3호선", "red", 1100
+            , "0530", "2300", 5);
+
+        신분당선.addSection(강남역, 양재역, 9, 2);
+        이호선.addSection(교대역, 강남역, 8, 3);
+        이호선.addSection(강남역, 삼성역, 8, 3);
+        삼호선.addSection(교대역, 남부터미널역, 6, 5);
+        삼호선.addSection(남부터미널역, 양재역, 6, 5);
+        삼호선.addSection(양재역, 대치역, 2, 5);
     }
 
     private Station createStation(long id, String name) {
